% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cancel.order}
\alias{cancel.order}
\title{Cancel an order}
\usage{
cancel.order(c, file = "~/.kraken_keys.txt")
}
\arguments{
\item{c}{The OID of the order. you can retreive it with \code{open.orders}}
}
\description{
Cancel an order
}
