% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.asks}
\alias{get.asks}
\title{Retrieve asks info from the last order book}
\usage{
get.asks(nn = 10)
}
\arguments{
\item{nn=10}{number of asks to show}
}
\description{
It only displays order book info
}
\keyword{bct}
\keyword{trades}
