% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.bids}
\alias{get.bids}
\title{Retrieve bids info from the last order book}
\usage{
get.bids(nn = 10)
}
\arguments{
\item{nn=10}{number of bids to show}
}
\description{
It only displays order book info
}
\keyword{bct}
\keyword{trades}
