% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.history}
\alias{get.history}
\title{Give a situation by a data.frame of [open,high,low,close] every t minutes since id}
\usage{
get.history(t = 1, id = 0)
}
\arguments{
\item{t}{Number of minutes between open and close. Possible numbers are 1 (default), 5, 15, 30, 60, 240 (4h), 1440 (1D), 10080 (1W), 21600 (15D). Default \code{t=1}}

\item{id}{The time in second from "1970-01-01" since you need the values. Default \code{id=0} in order to have the maximum amount possible of data}
}
\value{
A data frame with columns [time,open,high,low,close,vol]
}
\description{
Give a situation by a data.frame of [open,high,low,close] every t minutes since id
}
