% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/get.order.book.R
\name{get.order.book}
\alias{get.order.book}
\title{Retrieve order book informations}
\usage{
get.order.book()

get.order.book()
}
\description{
It only displays order book info

It only displays order book info
}
\keyword{bct}
\keyword{trades}
