% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/get.ticker.R
\name{get.ticker}
\alias{get.ticker}
\title{Retrieve ticker information}
\usage{
get.ticker()

get.ticker()
}
\description{
It only displays current highest bid and lowest ask

It only displays current highest bid and lowest ask
}
\keyword{bct}
\keyword{ticker}
