% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.order.book}
\alias{plot.order.book}
\title{Plot an order book}
\usage{
\method{plot}{order.book}(ob = get.order.book())
}
\arguments{
\item{ob}{the orderbook}
}
\description{
Plots an orderbook, default the last one available
}
\keyword{bct}
\keyword{trades}
