% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sell}
\alias{sell}
\title{Sell an amount of btc at a given price}
\usage{
sell(amount, price, file = "~/.kraken_keys.txt")
}
\arguments{
\item{amount}{Number of btc needed; min \code{0.01}}

\item{price}{Price per btc,  max precision \code{10^-1}}
}
\description{
Sell an amount of btc at a given price
}
